requirements.txt
cffi==1.14.3
Pillow==8.0.1
pycparser==2.20
pyheif==0.5.1
